requirements.txt
anthropic>=0.40.0
PyYAML>=6.0
rich>=13.0
prompt_toolkit>=3.0
sentence-transformers>=5.0
sqlite-vec>=0.1.6
python-dotenv>=1.0
beautifulsoup4>=4.12
html2text>=2024.2
